........
  r27800 | bugman | 2015-03-10 10:05:25 +0100 (Tue, 10 Mar 2015) | 7 lines
  
  Fix for the lib.sequence.check_sequence checking object.
  
  Although rarely used, the check for the spin number was incorrect and half of the checks were
  instead for the residue number.  This is a classic copy and paste error where the residue name and
  number checks were copied but not completely converted to spin name and numbers.
........
  r27801 | bugman | 2015-03-10 10:13:33 +0100 (Tue, 10 Mar 2015) | 5 lines
  
  Added the raise_flag argument to the lib.sequence.read_spin_data() function.
  
  This is to allow the missing data RelaxError to be deactivated.
........
  r27802 | bugman | 2015-03-10 10:15:35 +0100 (Tue, 10 Mar 2015) | 7 lines
  
  Modified the spectrum.read_intensities user function backend to be more robust.
  
  This affects the generic formatted peak lists, via the lib.spectrum.peak_list.intensity_generic()
  function.  The peak list reading will now continue reading the file after corrupted lines have been
  encountered.
........
  r27803 | bugman | 2015-03-10 18:00:53 +0100 (Tue, 10 Mar 2015) | 7 lines
  
  Python 3 improvement for the rdc.corr_plot and pcs.corr_plot user functions.
  
  The world view is now set in floating point numbers.  In Python 2, the math.ceil() and math.floor()
  functions return floats, whereas in Python 3 these functions return integers.  The behaviour is now
  consistent in both Python versions, fixing a few system tests.
........
  r27804 | bugman | 2015-03-10 18:36:55 +0100 (Tue, 10 Mar 2015) | 7 lines
  
  Improvements for the pcs.structural_noise user function.
  
  The check for the presence of PCS data for points to skip now includes checking for PCS values of
  None.  And the output Grace file now also includes the spin ID string as a string or comment value
  which can be displayed in the plot when desired.
........
  r27805 | bugman | 2015-03-10 18:58:41 +0100 (Tue, 10 Mar 2015) | 9 lines
  
  Modified the internal formatting of the data section of the Grace 2D graph files.
  
  This affects the lib.plotting.grace.write_xy_data() function.  The formatting is now more
  consistent, with the X value now set to a fixed number of decimal places, and hence will no longer
  change between Python 2 and 3.  The data is now all right justified as well, for easier reading.
  
  All affected system tests have been updated for the new format.
........
  r27807 | bugman | 2015-03-13 11:39:16 +0100 (Fri, 13 Mar 2015) | 3 lines
  
  Epydoc documentation fix for the lib.structure.pdb_write_handle_atom_name() function.
........
  r27816 | bugman | 2015-03-13 15:42:37 +0100 (Fri, 13 Mar 2015) | 6 lines
  
  Backported the relax 3.3.7 CHANGES file changes to trunk.
  
  The command used was:
  svn merge -r27810:r27811 svn+ssh://bugman@svn.gna.org/svn/relax/tags/3.3.7 .
........
  r27818 | bugman | 2015-03-13 17:55:09 +0100 (Fri, 13 Mar 2015) | 3 lines
  
  Spacing fixes in the CHANGES document.
........
  r27819 | bugman | 2015-03-17 11:12:55 +0100 (Tue, 17 Mar 2015) | 6 lines
  
  Fix for the pipe_control.reset.reset() function when resetting the GUI in non-standard contexts.
  
  This is mainly for debugging scripts when simulating a GUI and hence the GUI reset() method does not
  exist.
........
  r27820 | bugman | 2015-03-17 12:21:57 +0100 (Tue, 17 Mar 2015) | 7 lines
  
  Created a GUI memory management debugging script for the align_tensor.init user function.
  
  This repetitively calls the reset, pipe.create and align_tensor.init user functions, and opening the
  GUI element for setting alignment tensor elements (the Sequence window).  The pympler muppy_log file
  shows no memory leaks for these user functions on Linux systems.
........
  r27822 | bugman | 2015-04-02 09:14:10 +0200 (Thu, 02 Apr 2015) | 10 lines
  
  Resized all fixed-sized GUI wizards to fit on 1024x768 pixel wide displays.
  
  The problem was reported by Lora Picton in the thread starting at
  http://thread.gmane.org/gmane.science.nmr.relax.user/1813 .
  
  Both the spin loading wizard of the spin viewer window and the relaxation data loading wizard used
  currently in the model-free analysis tab and BMRB export page were fixed.  These both had the
  y-dimension set to 800 pixels, hence parts of the window would be out of view.
........
  r27827 | bugman | 2015-04-02 09:41:52 +0200 (Thu, 02 Apr 2015) | 6 lines
  
  Backported the relax 3.3.8 CHANGES file changes to trunk.
  
  The command used was:
  svn merge -r27825:r27826 svn+ssh://bugman@svn.gna.org/svn/relax/tags/3.3.8 .
........

+++ pipe_control/rdc.py
@@ -548,8 +548,8 @@ def corr_plot(format=None, title=None, subtitle=None, file=None, dir=None, force
     size = len(data)
 
     # Round the data limits.
-    max_rdc = ceil(max_rdc)
-    min_rdc = floor(min_rdc)
+    max_rdc = float(ceil(max_rdc))
+    min_rdc = float(floor(min_rdc))
 
     # Only one data set.
     data = [data]
